Feature - recall tasks
* added recall and discard task content to all task types
* removed withdraw component
* refactored to remove repetition
