+++ pages/task/read/routers/view.js
@@ -44,11 +44,6 @@ module.exports = () => {
 
   app.use((req, res, next) => {
     req.breadcrumb('task.read');
-    const action = get(req, 'query.action');
-    if (action === 'withdraw') {
-      set(req, `session.form[${req.task.id}].values`, { status: 'withdrawn-by-applicant' });
-      return res.redirect(req.buildRoute('task.confirm', req.params));
-    }
     next();
   });
