+++ pages/task/read/views/components/deadline.jsx
@@ -1,8 +1,8 @@
 import React, { Component, Fragment } from 'react';
 import { connect } from 'react-redux';
 import { Link, Snippet } from '@asl/components';
-import { dateFormat } from '../../../../constants';
-import { formatDate } from '../../../../lib/utils';
+import { dateFormat } from '../../../../../constants';
+import { formatDate } from '../../../../../lib/utils';
 
 class Deadline extends Component {
   render() {
