+++ pages/task/read/views/models/profile.jsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import omit from 'lodash/omit';
+import {
+  Snippet,
+  StickyNavAnchor,
+  Diff
+} from '@asl/components';
+import userSchema from '../../../../user/update/schema';
+import { dateFormat } from '../../../../../constants';
+import { formatDate } from '../../../../../lib/utils';
+
+const formatters = {
+  dob: {
+    format: date => formatDate(date, dateFormat.short)
+  }
+};
+
+export default function Profile({ task, values, schema }) {
+  return (
+    <StickyNavAnchor id="changes">
+      <h2><Snippet>sticky-nav.changes</Snippet></h2>
+      <Diff values={task.data.data} model={values} schema={omit(userSchema, 'comments')} formatters={formatters} />
+    </StickyNavAnchor>
+  );
+}
