+++ pages/task/schema/view.js
@@ -5,10 +5,6 @@ module.exports = task => {
   const content = getContent(task);
 
   const options = task.nextSteps.map(option => {
-    if (option.id === 'withdrawn-by-applicant') {
-      task.canBeWithdrawn = true;
-    }
-
     return {
       value: option.id,
       label: get(content, `status.${option.id}.action.${task.type}`, get(content, `status.${option.id}.action`)),
@@ -19,7 +15,7 @@ module.exports = task => {
   return {
     status: {
       inputType: 'radioGroup',
-      options: options.filter(option => option.value !== 'withdrawn-by-applicant'),
+      options,
       nullValue: [],
       label: get(content, `fields.status.${task.status}.${task.type}`, get(content, `fields.status.${task.status}`, get(content, 'fields.status.label'))),
       validate: [
